requirements.txt
streamlit
pandas
plotly
mysql-connector-pythoncd 